tests/test_routes_projects.py
"""
Tests for Project Routes
Tests the project routes once they are migrated to the new blueprint architecture

NOTE: These tests are currently SKIPPED because the legacy app.py routes
have not yet been migrated to blueprints. Once the migration is complete,
these tests will validate the new blueprint-based project routes.
"""

import pytest
from flask import json
from typing import Dict, Any


pytestmark = pytest.mark.skip(reason="Legacy app.py routes not yet migrated to blueprints")


class TestProjectRoutes:
    """Test suite for project-related routes"""

    def test_get_projects_success(self, client, mock_db_cursor):
        """Test GET /api/projects returns project list successfully"""
        # Configure mock to return sample projects
        sample_projects = [
            {
                'project_id': 'f47ac10b-58cc-4372-a567-0e02b2c3d479',
                'project_name': 'Test Project 1',
                'project_number': 'PRJ-001',
                'client_name': 'Test Client',
                'description': 'Sample project for testing',
                'quality_score': 0.85,
                'tags': '{}',
                'attributes': '{}'
            },
            {
                'project_id': 'a1b2c3d4-58cc-4372-a567-0e02b2c3d480',
                'project_name': 'Test Project 2',
                'project_number': 'PRJ-002',
                'client_name': 'Another Client',
                'description': 'Another sample project',
                'quality_score': 0.75,
                'tags': '{}',
                'attributes': '{}'
            }
        ]
        mock_db_cursor.fetchall.return_value = sample_projects

        # Make request
        response = client.get('/api/projects')

        # Assert response
        assert response.status_code == 200
        data = json.loads(response.data)
        assert 'projects' in data
        assert isinstance(data['projects'], list)

    def test_get_projects_empty(self, client, mock_db_cursor):
        """Test GET /api/projects returns empty list when no projects exist"""
        # Configure mock to return empty list
        mock_db_cursor.fetchall.return_value = []

        # Make request
        response = client.get('/api/projects')

        # Assert response
        assert response.status_code == 200
        data = json.loads(response.data)
        assert 'projects' in data
        assert data['projects'] == []

    def test_create_project_success(self, client, mock_db_cursor, mock_db_connection):
        """Test POST /api/projects creates a new project successfully"""
        # Configure mock cursor to return created project data
        created_project = (
            'f47ac10b-58cc-4372-a567-0e02b2c3d479',  # project_id
            'New Test Project',                       # project_name
            None,                                     # client_id
            'New Client',                             # client_name
            'PRJ-NEW-001',                            # project_number
            'system-id-123',                          # default_coordinate_system_id
            '2025-11-18T10:00:00'                     # created_at
        )
        mock_db_cursor.fetchone.return_value = created_project

        # Prepare request data
        project_data = {
            'project_name': 'New Test Project',
            'client_name': 'New Client',
            'project_number': 'PRJ-NEW-001',
            'description': 'A newly created test project'
        }

        # Make request
        response = client.post(
            '/api/projects',
            data=json.dumps(project_data),
            content_type='application/json'
        )

        # Assert response
        assert response.status_code == 200
        data = json.loads(response.data)
        assert data['project_name'] == 'New Test Project'
        assert data['client_name'] == 'New Client'
        assert data['project_number'] == 'PRJ-NEW-001'
        assert 'project_id' in data

    def test_create_project_missing_name(self, client):
        """Test POST /api/projects fails when project_name is missing"""
        # Prepare request data without project_name
        project_data = {
            'client_name': 'Test Client',
            'description': 'Missing project name'
        }

        # Make request
        response = client.post(
            '/api/projects',
            data=json.dumps(project_data),
            content_type='application/json'
        )

        # Assert response
        assert response.status_code == 400
        data = json.loads(response.data)
        assert 'error' in data
        assert 'project_name is required' in data['error']

    def test_create_project_invalid_json(self, client):
        """Test POST /api/projects fails with invalid JSON"""
        # Make request with invalid JSON
        response = client.post(
            '/api/projects',
            data='invalid json{',
            content_type='application/json'
        )

        # Assert response (400 for invalid JSON or no data)
        assert response.status_code in [400, 500]

    def test_get_project_by_id_success(self, client, mock_db_cursor):
        """Test GET /api/projects/<id> returns specific project"""
        # Configure mock to return project
        project = {
            'project_id': 'f47ac10b-58cc-4372-a567-0e02b2c3d479',
            'project_name': 'Test Project 1',
            'project_number': 'PRJ-001',
            'client_name': 'Test Client',
            'description': 'Sample project for testing',
            'quality_score': 0.85
        }
        mock_db_cursor.fetchone.return_value = project
        mock_db_cursor.fetchall.return_value = [project]

        # Make request
        response = client.get('/api/projects/f47ac10b-58cc-4372-a567-0e02b2c3d479')

        # Assert response
        assert response.status_code == 200
        data = json.loads(response.data)
        # The response could be the project directly or wrapped
        if 'project_id' in data:
            assert data['project_id'] == 'f47ac10b-58cc-4372-a567-0e02b2c3d479'
            assert data['project_name'] == 'Test Project 1'

    def test_get_project_by_id_not_found(self, client, mock_db_cursor):
        """Test GET /api/projects/<id> handles missing project"""
        # Configure mock to return None
        mock_db_cursor.fetchone.return_value = None
        mock_db_cursor.fetchall.return_value = []

        # Make request
        response = client.get('/api/projects/nonexistent-id')

        # Assert response (could be 404 or 200 with error)
        assert response.status_code in [200, 404, 500]

    def test_delete_project_success(self, client, mock_db_cursor):
        """Test DELETE /api/projects/<id> deletes project successfully"""
        # Configure mock
        mock_db_cursor.rowcount = 1

        # Make request
        response = client.delete('/api/projects/f47ac10b-58cc-4372-a567-0e02b2c3d479')

        # Assert response
        assert response.status_code == 200
        data = json.loads(response.data)
        assert 'success' in data
        assert data['success'] is True


class TestProjectRoutesIntegration:
    """Integration tests for project routes (marked as integration)"""

    @pytest.mark.integration
    def test_full_project_lifecycle(self, client, mock_db_cursor):
        """Test creating, retrieving, and deleting a project"""
        # This test would normally use real database
        # For now, using mocks to demonstrate the pattern

        # 1. Create project
        created_project = (
            'lifecycle-test-id',
            'Lifecycle Test Project',
            None,
            'Test Client',
            'LT-001',
            None,
            '2025-11-18T10:00:00'
        )
        mock_db_cursor.fetchone.return_value = created_project

        create_response = client.post(
            '/api/projects',
            data=json.dumps({
                'project_name': 'Lifecycle Test Project',
                'client_name': 'Test Client',
                'project_number': 'LT-001'
            }),
            content_type='application/json'
        )
        assert create_response.status_code == 200

        # 2. Retrieve project
        mock_db_cursor.fetchone.return_value = {
            'project_id': 'lifecycle-test-id',
            'project_name': 'Lifecycle Test Project',
            'project_number': 'LT-001'
        }
        get_response = client.get('/api/projects/lifecycle-test-id')
        assert get_response.status_code == 200

        # 3. Delete project
        delete_response = client.delete('/api/projects/lifecycle-test-id')
        assert delete_response.status_code == 200
