requirements.txt
affine==2.4.0
annotated-types==0.7.0
anyio==4.11.0
arcgis2geojson==3.1.1
attrs==25.4.0
bidict==0.23.1
blinker==1.9.0
brotli==1.2.0
cachelib==0.13.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-plugins==1.1.1.2
cligj==0.7.2
contourpy==1.3.3
cssselect2==0.8.0
cycler==0.12.1
distro==1.9.0
dnspython==2.8.0
et_xmlfile==2.0.0
eventlet==0.40.3
ezdxf==1.4.3
fiona==1.10.1
Flask==3.1.2
Flask-Caching==2.3.1
flask-cors==6.0.1
Flask-SocketIO==5.5.1
fonttools==4.60.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
kiwisolver==1.4.9
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
networkx==3.5
numpy==2.3.4
openai==2.6.1
openpyxl==3.1.5
OWSLib==0.35.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydyf==0.11.0
pyparsing==3.2.5
pyphen==0.17.2
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.3
python-socketio==5.14.3
pytz==2025.2
PyYAML==6.0.3
rasterio==1.4.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
shapely==2.1.2
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
wsproto==1.3.1
zopfli==0.4.0
